Split conversation component into smaller child components to avoid maintenance issues
